Upgrading Tardis version, Modify Rbum kernel model (in progress)

+++ basic/src/rbum/rbum_config.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(default)]
 pub struct RbumConfig {
-    pub set_cate_sys_code_node_len: u8,
+    pub set_cate_sys_code_node_len: usize,
 }
 
 impl Default for RbumConfig {
